+++ dm_control/mujoco/engine_test.py
@@ -532,7 +532,7 @@ class MujocoEngineTest(parameterized.TestCase):
     """
     physics = engine.Physics.from_xml_string(xml)
     spec = engine.action_spec(physics)
-    self.assertEqual(np.float, spec.dtype)
+    self.assertEqual(float, spec.dtype)
     np.testing.assert_array_equal(spec.minimum, [-constants.mjMAXVAL, -1.0])
     np.testing.assert_array_equal(spec.maximum, [constants.mjMAXVAL, 2.0])
